+++ app/build.gradle
@@ -175,9 +175,6 @@ dependencies {
     // HockeyApp
     implementation libs.hockeyApp
 
-    // CircleImageView
-    implementation libs.circleImageView
-
     // Unit testing
     testImplementation test.junit
     testImplementation test.mockito
