+++ app/src/main/AndroidManifest.xml
@@ -42,7 +42,7 @@
             android:screenOrientation="portrait" />
 
         <activity
-            android:name=".feature.client.home.activity.ClientHomeActivity"
+            android:name=".feature.client.home.ClientHomeActivity"
             android:screenOrientation="portrait" />
 
         <meta-data
