+++ app/src/main/kotlin/co/netguru/baby/monitor/client/feature/client/home/dashboard/ClientDashboardFragment.kt
@@ -1,4 +1,4 @@
-package co.netguru.baby.monitor.client.feature.client.home.fragment
+package co.netguru.baby.monitor.client.feature.client.home.dashboard
 
 import android.os.Bundle
 import android.support.v4.app.Fragment
@@ -7,10 +7,9 @@ import android.view.View
 import android.view.ViewGroup
 import co.netguru.baby.monitor.client.R
 
-class ClientSettingsFragment: Fragment() {
+class ClientDashboardFragment : Fragment() {
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        return inflater.inflate(R.layout.fragment_client_settings, container, false)
+        return inflater.inflate(R.layout.fragment_client_dashboard, container, false)
     }
-
 }
