+++ app/src/main/kotlin/co/netguru/baby/monitor/client/feature/client/home/log/ClientActivityLogFragment.kt
@@ -1,4 +1,4 @@
-package co.netguru.baby.monitor.client.feature.client.home.fragment
+package co.netguru.baby.monitor.client.feature.client.home.log
 
 import android.os.Bundle
 import android.support.v4.app.Fragment
@@ -9,7 +9,9 @@ import co.netguru.baby.monitor.client.R
 
 class ClientActivityLogFragment : Fragment() {
 
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?
+    ): View? {
         return inflater.inflate(R.layout.fragment_client_activity_log, container, false)
     }
 }
