+++ app/src/main/kotlin/co/netguru/baby/monitor/client/feature/client/home/lullabies/ClientLullabiesFragment.kt
@@ -1,4 +1,4 @@
-package co.netguru.baby.monitor.client.feature.client.home.fragment
+package co.netguru.baby.monitor.client.feature.client.home.lullabies
 
 import android.os.Bundle
 import android.support.v4.app.Fragment
@@ -12,5 +12,4 @@ class ClientLullabiesFragment : Fragment() {
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
         return inflater.inflate(R.layout.fragment_client_lullabies, container, false)
     }
-
 }
