+++ app/src/main/kotlin/co/netguru/baby/monitor/client/feature/welcome/WelcomeFragment.kt
@@ -6,7 +6,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import co.netguru.baby.monitor.client.R
-import co.netguru.baby.monitor.client.feature.client.home.activity.ClientHomeActivity
+import co.netguru.baby.monitor.client.feature.client.home.ClientHomeActivity
 import co.netguru.baby.monitor.client.feature.server.ServerActivity
 import kotlinx.android.synthetic.main.fragment_welcome.*
 import org.jetbrains.anko.support.v4.startActivity
