+++ app/src/main/res/layout/activity_client_home.xml
@@ -24,7 +24,5 @@
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/clientHomeFrameLayout"
-        app:layout_constraintVertical_bias="1"
         app:menu="@menu/home_menu" />
-
 </android.support.constraint.ConstraintLayout>
