+++ build.gradle
@@ -2,7 +2,7 @@
 buildscript {
 
     ext {
-        kotlinVersion = '1.2.60'
+        kotlinVersion = '1.2.71'
     }
     repositories {
         jcenter()
@@ -13,7 +13,7 @@ buildscript {
     }
 
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.1.4'
+        classpath 'com.android.tools.build:gradle:3.2.0'
         classpath "com.getkeepsafe.dexcount:dexcount-gradle-plugin:0.8.2"
         classpath "com.frogermcs.androiddevmetrics:androiddevmetrics-plugin:0.6"
         classpath 'com.vanniktech:gradle-android-junit-jacoco-plugin:0.11.0'
