+++ buildsystem/dependencies.gradle
@@ -55,7 +55,6 @@ ext {
             paperParcel             : "nz.bradcampbell:paperparcel:$paperParcelVersion",
             libVLC                  : "de.mrmaffen:libvlc-android:2.1.12@aar",
             lifecycleExtensions     : "android.arch.lifecycle:extensions:$lifecycle_version",
-            circleImageView         : "de.hdodenhof:circleimageview:2.2.0"
     ]
 
     proc = [
